Added form on register.html and tied it in with Authctrl
Tested on Chrome Safari and Firefox, expect it to work. Some ancillary
errors on firefox, but still works.

In authCtrl.js added a console.log to log the $scope.signup function

In register.html added the forms to take in the values and tied them to
angular. On lines 5 and 6 added go back to login

In styles.css added a class for the input fields - .register