Added routing and controller for all issue pages. Pruned some files.
Added all the routes for the upcoming issue pages.

In home.html, which was moved to the issueViews folder, fixed all the
paths for the image files, and added the hrefs for the URLS.

In navigatorCtrl.js added the navigatorCtrl.

in app.js added all the routing information for each of the issueViews.

In app.js removed myApp.home and myApp.login and myApp.register. Just
unnecessary.

Removed files home.js, login.js and register.js. Unecessary. The
controllers were changed to the same authCtrl, and the routing
information was all there, so did not need these extra modules.

In authCtrl.js on lines 36-52 added the logic to decide what type of
account the user wants to register for the transition from initial
login page to register page

In login.html added the corresponding changes to each image to receive
type of account value

in index.html added issueViews/navigatorCtrl.js on line 37, removed
login.js, register.js and home.js from lines 37-39