Added question page and attached some parts to angular
Tested in Firefox Chrome and Safari

In index.html added questionsctrl.js script

In app.js added the routing info for /questions

In navigator ctrl fixed a console.log

Added questions.html, fixed the buttons to angular with ng-value and
ng-click. Still need to tie uploading these values to the next question
button as well as make the add comments button work.

In questionCtrl.js added the questions controller and a
$scope.answerQuestion that will submit the info, and a
$scope.importance thing that is just a test function for some of the
fields