Made the home page interest categories partly response
THIS ONLY WORKS ON CHROME. IN ITS OWN BRANCH FOR THE TIME BEING.

In home.html, divided all the interest hrefs into <div
class=“flex-item”> and put them in <div class=“flex-container”>, so the
ROWS reorganize themselves responsively to the width of the browser.
STILL need to make the actual image sizes responsive… maybe this change
is redundant when that happens, too.

in category.html, added class “background image” to div with ng-repeat
in it, so it loads the background images. This will work when sized
correctly, but we will need .PNG images so that the opacity doesn’t
give us an ugly gray color.

In index.html moved things around, made it read a little better in the
<head>

In styles added .backgroundImage on line 90 for use with
categories.html, and commented it out for now.

Added flex-container on line 259 and flex-item on line 274 for use with
home.html to arrange all the categories in rows.